Extract all the various subcommands into their own.
Vaugely nods in the direction of bringing ce_install (optionally)
under the same system (i.e. `ce install`).
